+++ src/components/manage/Add/Add.jsx
@@ -180,6 +180,10 @@ class Add extends Component {
   render() {
     if (this.props.schemaRequest.loaded) {
       const visual = hasBlocksData(this.props.schema.properties);
+      const initialBlocks =
+        this.props.schema.properties.blocks?.default || this.initialBlocks;
+      const initialBlocksLayout = this.props.schema.properties.blocks_layout
+        ?.default || { items: this.initialBlocksLayout };
 
       return (
         <div id="page-add">
@@ -192,11 +196,10 @@ class Add extends Component {
             ref={this.form}
             schema={this.props.schema}
             formData={{
-              [getBlocksFieldname(this.props.schema.properties)]: this
-                .initialBlocks,
-              [getBlocksLayoutFieldname(this.props.schema.properties)]: {
-                items: this.initialBlocksLayout,
-              },
+              [getBlocksFieldname(this.props.schema.properties)]: initialBlocks,
+              [getBlocksLayoutFieldname(
+                this.props.schema.properties,
+              )]: initialBlocksLayout,
             }}
             onSubmit={this.onSubmit}
             hideActions